+++ native/misc.h
@@ -1,7 +1,7 @@
 void primitive_exit(void);
 void primitive_os_env(void);
 void primitive_eq(void);
-long long current_millis(void);
+int64_t current_millis(void);
 void primitive_millis(void);
 void primitive_init_random(void);
 void primitive_random_int(void);
